==> 5_statistical_analyses.ipynb <==
### Statistical Analyses
#### List of tasks accomplished in this Jupyter Notebook:
- Print out the n values for each treatment for reference
- Compare physiological (length) differences between males, females, and fed vs starved larvae
- Compare larval activity to time of day
- Compare differences between fed and starved larvae in behavioral metrics
- Calculate the preference value for each stimulus for both fed and starved larvae
- Compare behavioral metrics for neutral, aversive, and appetitive cues
- Compare the effect of larval presence on dye diffusion over time
- Assess fit of the regression line used to fit distance to concentration for simulations
- Adjust all tests using the Holm-Bonferroni correction

R.version # print R version for reference

setwd("/Users/ghost/Documents/Gitrepos/aedes-aegypti-2020/2019_data")
data <- read.csv('./data/trajectories/summary/cleaned_animal_analyses.csv')

data$A_starved <- as.factor(data$A_starved)
data$A_treatment_odor <- as.factor(data$A_treatment_odor)
data$A_sex <- as.factor(data$A_sex)

method_n <- "holm"

# Print out the n values for each treatment for reference

cat("Starved larvae:")
sub <- subset(data, A_starved=="1day")
data.frame(table(sub$A_treatment_odor))

sub <- subset(data, A_starved=="1day")
data.frame(table(sub$A_sex))

cat("Fed larvae:")
sub <- subset(data, A_starved=="no")
data.frame(table(sub$A_treatment_odor))

sub <- subset(data, A_starved=="no")
data.frame(table(sub$A_sex))

# Test for normality of larval size data to be compared. 
# If distributions are not normal, a nonparametric test must be used. 

setwd("/Users/ghost/Documents/Gitrepos/aedes-aegypti-2020/2019_data")
data <- read.csv('./cleaned_animal_analyses.csv')

data$A_starved <- as.factor(data$A_starved)
data$A_treatment_odor <- as.factor(data$A_treatment_odor)
data$A_sex <- as.factor(data$A_sex)

"TRUE/FALSE: Data queried has a normal distribution"
"All tests must be TRUE to use statistical tests that assume normality"

sub <- subset(data, A_sex=="f")
sub1 <- subset(sub, A_starved=="1day")
sub2 <- subset(sub, A_starved=="no")
shapiro.test(sub1$A_larvae_length_mm)$p.value>0.05
shapiro.test(sub2$A_larvae_length_mm)$p.value>0.05

sub <- subset(data, A_sex=="m")
sub1 <- subset(sub, A_starved=="1day")
sub2 <- subset(sub, A_starved=="no")
shapiro.test(sub1$A_larvae_length_mm)$p.value>0.05
shapiro.test(sub2$A_larvae_length_mm)$p.value>0.05

# Compare physiological (length) differences between males, females, 
# and fed vs starved larvae

# starved females vs fed females
sub <- subset(data, A_sex=="f")
sub1 <- subset(sub, A_starved=="1day")
sub2 <- subset(sub, A_starved=="no")
resp <- wilcox.test(sub1$A_larvae_length_mm, sub2$A_larvae_length_mm)
c(mean(sub2$A_larvae_length_mm) - mean(sub1$A_larvae_length_mm), 'f fed - f starved mean')
p <- resp$p.value
t <- "f starved vs f fed size"

# starved males vs fed males
sub <- subset(data, A_sex=="m")
sub1 <- subset(sub, A_starved=="1day")
sub2 <- subset(sub, A_starved=="no")
resp <- wilcox.test(sub1$A_larvae_length_mm, sub2$A_larvae_length_mm)
c(mean(sub2$A_larvae_length_mm) - mean(sub1$A_larvae_length_mm), 'm fed - m starved mean')
p <- c(p, resp$p.value)
t <- c(t, "m starved vs m fed size")

# starved females vs starved males
sub <- subset(data, A_starved=="1day")
sub1 <- subset(sub, A_sex=="f")
sub2 <- subset(sub, A_sex=="m")
resp <- wilcox.test(sub1$A_larvae_length_mm, sub2$A_larvae_length_mm)
c(mean(sub1$A_larvae_length_mm) - mean(sub2$A_larvae_length_mm), 'f starved mean - m starved mean')
p <- c(p, resp$p.value)
t <- c(t, "m starved vs f starved size")

# fed females vs fed males
sub <- subset(data, A_starved=="no")
sub1 <- subset(sub, A_sex=="f")
sub2 <- subset(sub, A_sex=="m")
resp <- wilcox.test(sub1$A_larvae_length_mm, sub2$A_larvae_length_mm)
c(mean(sub1$A_larvae_length_mm) - mean(sub2$A_larvae_length_mm), 'f fed mean - m fed mean')
p <- c(p, resp$p.value)
t <- c(t, "f fed vs m fed size")

# Compare larval activity to time of day
resp <- lm(data$A_time_move~data$A_minutes_past_L)
r <- anova(resp)$'Pr(>F)'[1]
p <- c(p, r)
t <- c(t, "time moving by daylight regression")

resp <- lm(data$A_time_wall~data$A_minutes_past_L)
r <- anova(resp)$'Pr(>F)'[1]
p <- c(p, r)
t <- c(t, "time next to wall by daylight regression")

resp <- lm(data$A_median_speed~data$A_minutes_past_L)
r <- anova(resp)$'Pr(>F)'[1]
p <- c(p, r)
t <- c(t, "median speed by daylight regression")

# Compare differences between fed and starved larvae in behavioral metrics

sub1 <- subset(data, A_starved=="1day")
sub2 <- subset(data, A_starved=="no")
resp <- wilcox.test(sub1$A_time_move, sub2$A_time_move)
p <- c(p, resp$p.value)
t <- c(t, "time moving by starvation state")

sub1 <- subset(data, A_starved=="1day")
sub2 <- subset(data, A_starved=="no")
resp <- wilcox.test(sub1$A_time_wall, sub2$A_time_wall)
p <- c(p, resp$p.value)
t <- c(t, "time next to wall by starvation state")

cat("Number of tests ran:", length(p))
p <- p.adjust(p, method=method_n)
d <- data.frame(t, p, p<0.05, p<0.01, p<0.001, p<0.0001)
d

# Test for normality of behavior metrics to be compared. 
# If distributions are not normal, a nonparametric test must be used. 

data3 <- read.csv('./data/trajectories/summary/cleaned_animal_analyses_stimuli_groups.csv')
data3$stimulus <- as.factor(data3$stimulus)

treatments <- c("naive_100ul_left_food_05percent", "naive_100ul_left_food_extract", 
                "naive_100ul_left_yeastRNA_1gL", "naive_100ul_left_quinine_10mM",
                "naive_100ul_left_indole_100uM", "naive_100ul_left_o-cresol_100uM", 
                "naive_100ul_left_milliQ_water", "naive_100ul_left_indole_10mM",
                "naive_100ul_left_amino_acids", "naive_100ul_left_glucose_10gL")

a <- "Normal"
b <- "Normal"
c <- "Normal"
d <- "Normal"
e <- "Normal"
f <- "Normal"
g <- "Normal"

for (fed in c("1day", "no")){
    for (treatment in treatments){
        ss <- subset(data3, A_treatment_odor==treatment)
        ss <- subset(ss, A_starved==fed) 
        if (shapiro.test(ss$median_conc_diff)$p.value<0.05){
            a <- "Not normally distributed"
        }
        if (shapiro.test(ss$cd_move_diff)$p.value<0.05){
            b <- "Not normally distributed"
        }
        if (shapiro.test(ss$discovery_time_diff)$p.value<0.05){
            c <- "Not normally distributed"
        }
        if (shapiro.test(ss$c_speed_diff)$p.value<0.05){
            d <- "Not normally distributed"
        }
        if (shapiro.test(ss$cd_speed_diff)$p.value<0.05){
            e <- "Not normally distributed"
        }
        if (shapiro.test(ss$c_turn_diff)$p.value<0.05){
            f <- "Not normally distributed"
        }
        if (shapiro.test(ss$cd_turn_diff)$p.value<0.05){
            g <- "Not normally distributed"
        }}}

"Check if data queried has a normal distribution"
"All tests must be NORMAL to use statistical tests that assume normality"

a
b
c
d
e
f
g

# Calculate the preference value for each stimulus for both fed and starved larvae

data <- read.csv('./data/trajectories/summary/cleaned_animal_analyses_stimuli_groups.csv')
data$stimulus <- as.factor(data$stimulus)

treatments <- c("naive_100ul_left_food_05percent", "naive_100ul_left_food_extract", 
                "naive_100ul_left_yeastRNA_1gL", "naive_100ul_left_quinine_10mM",
                "naive_100ul_left_indole_100uM", "naive_100ul_left_o-cresol_100uM", 
                "naive_100ul_left_milliQ_water", "naive_100ul_left_indole_10mM", 
                "naive_100ul_left_amino_acids", "naive_100ul_left_glucose_10gL")
p <- 'test'
t <- 'test'
for (fed in c("1day", "no")){
    for (treatment in treatments){
        ss <- subset(data, A_treatment_odor==treatment)
        ss <- subset(ss, A_starved==fed)
        resp <- t.test(ss$A_median_conc, ss$E_median_conc, paired=TRUE, alternative="two.sided")
        p <- c(p, resp$p.value)
        t <- c(t, paste(treatment, fed, "paired test of median concentration", sep=" "))}}

# Compare behavioral metrics for neutral, aversive, and appetitive cues
data3 <- read.csv('./data/trajectories/summary/cleaned_animal_analyses_stimuli_groups.csv')
data3$stimulus <- as.factor(data3$stimulus)

resp <- kruskal.test(data3$median_conc_diff~data3$stimulus)
p <- c(p, resp$p.value)
t <- c(t, "median concentration by odor")

resp <- kruskal.test(data3$cd_move_diff~data3$stimulus)
p <- c(p, resp$p.value)
t <- c(t, "cd_move_diff by odor")

resp <- kruskal.test(data3$discovery_time_diff~data3$stimulus)
p <- c(p, resp$p.value)
t <- c(t, "discovery_time by odor")

resp <- kruskal.test(data3$c_speed_diff~data3$stimulus)
p <- c(p, resp$p.value)
t <- c(t, "c_speed_diff by odor")

resp <- kruskal.test(data3$cd_speed_diff~data3$stimulus)
p <- c(p, resp$p.value)
t <- c(t, "cd_speed_diff by odor")

resp <- kruskal.test(data3$c_turn_diff~data3$stimulus)
p <- c(p, resp$p.value)
t <- c(t, "c_turn_diff by odor")

resp <- kruskal.test(data3$cd_turn_diff~data3$stimulus)
p <- c(p, resp$p.value)
t <- c(t, "cd_turn_diff by odor")

# remove placeholder values and print out p values
p <- p[-1]; 
t <- t[-1];
cat("Number of tests ran:", length(p))

p <- p.adjust(p, method=method_n)
d <- data.frame(t, p, p<0.05, p<0.01, p<0.001, p<0.0001)
d

- Assess fit of the regression line used to fit distance to concentration for simulations

# Compare the effect of larval presence on dye diffusion over time

# Mann-Whitney Wilcoxon test for unpaired data
data <- read.csv("./larvae_no_larvae_comparison.csv")
data <- subset(data, data$time == 0)
sub1 <- subset(data, larva_presence=="no_larvae")
sub2 <- subset(data, larva_presence=="larvae")
resp <- wilcox.test(sub1$perc_over_50, sub2$perc_over_50)
pp <- resp$p.value
tt <- "Dye distribution at time 0"

# Linear regression for all other data points
data <- read.csv("./larvae_no_larvae_comparison.csv")
data$larva_presence = as.factor(data$larva_presence)
data <- subset(data, data$time > 0)
resp <- lm(data$perc_over_50 ~ data$time+data$larva_presence)
summary(resp)

pp <- c(pp, 2e-16)
tt <- c(tt, "regression larval presence time concentration: intercept")
pp <- c(pp, 0.000777)
tt <- c(tt, "regression larval presence time concentration: time")
pp <- c(pp, 7.23e-12)
tt <- c(tt, "regression larval presence time concentration: presence")

pp <- p.adjust(pp, method=method_n)
dd <- data.frame(tt, pp, pp<0.05)
dd

# See fit of regression line. The regression itself was done in Python. 

data2 <- read.csv("./data/fluorescein/distance_concentration_map_fitted.csv")
resp <- lm(data2$ln_conc~data2$distance_mm)
           
summary(resp)
r <- anova(resp)$'Pr(>F)'[1]
r
