==> 6_combine_trajectory_data_for_modeling.ipynb <==
### Combine trajectory data
#### List of tasks accomplished in this Jupyter Notebook:
- Output 4 dataframe combining all animal trajectories: Fed animals acclimation phase, Fed animals experiment phase, Starved animals acclimation phase, and Starved animals experiment phase

import numpy as np
import pandas as pd
import eleanor_constants as EL

df = pd.read_csv("./data/experiment_IDs/cleaned_static_data.csv")

for val in ["acclimate", "experiment"]:
    for food, tag in EL.fed.items():
        df_food = df[df['starved'] == tag]
        master_df = pd.DataFrame()

        for index, row in df_food.iterrows():
            animal = row["animal_ID"]
            readname = "./data/trajectories/video_calculations/"+animal+"-"+val+".csv"
            temp = pd.read_csv(readname)
            temp["animal_ID"] = animal
            temp["treatment_odor"] = row["treatment_odor"]
            master_df = pd.concat([master_df, temp], sort=False)

        master_df.drop(["interpolated", "manual_tracker_fix", "objid", "pixel_height", "pixel_width",
                        "measurement_x", "measurement_y", "position_x", "position_y", "bin_ID", 
                        "turn", "larvae_length_mm", 
                        "pos_x_mm", "pos_y_mm"],
                        axis=1, inplace=True)
        master_df.to_csv("./data/trajectories/summary/modeling_"+\
                         food+"_"+val+"_all_animals.csv", index=None)

print("--- All files finished ---")
